files created, initial tests passed, upload and db working
