+++ qa_app.py
@@ -1,0 +1,89 @@
+import os
+from transformers import AutoTokenizer, AutoModelForSeq2SeqLM 
+from transformers import pipeline
+import torch 
+from langchain.document_loaders import PyPDFLoader
+from langchain.text_splitter import RecursiveCharacterTextSplitter 
+from langchain.embeddings import SentenceTransformerEmbeddings 
+from langchain.vectorstores import Chroma 
+from langchain.llms import HuggingFacePipeline
+from langchain.chains import RetrievalQA 
+
+
+device = torch.device('cpu')
+
+checkpoint = "LaMini-T5-738M"
+
+print(f"Checkpoint path: {checkpoint}")  
+tokenizer = AutoTokenizer.from_pretrained(checkpoint)
+base_model = AutoModelForSeq2SeqLM.from_pretrained(
+    checkpoint,
+    torch_dtype=torch.float32
+)
+
+def data_ingestion2():
+    for root, dirs, files in os.walk("docs"):
+        for file in files:
+            if file.endswith(".pdf"):
+                print(file)
+                loader = PyPDFLoader(os.path.join(root, file))
+    documents = loader.load()
+    print("splitting into chunks")
+    text_splitter = RecursiveCharacterTextSplitter(chunk_size=500, chunk_overlap=100)
+    texts = text_splitter.split_documents(documents)
+    #create embeddings here
+    print("Loading sentence transformers model")
+    embeddings = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")
+    #create vector store here
+    print(f"Creating embeddings. May take some minutes...")
+
+    db = Chroma.from_documents(texts, embeddings, persist_directory="vector_db_main")
+    # print(type(db))
+    # query = "What is the EconVisor's motivation is deeply rooted in?"
+    # docs = db.similarity_search(query)
+    # print(docs[0].page_content)
+
+    print(f"Ingestion complete! You can now run privateGPT.py to query your documents")
+
+def llm_pipeline():
+    pipe = pipeline(
+        'text2text-generation',
+        model = base_model,
+        tokenizer = tokenizer,
+        max_length = 256,
+        do_sample = True,
+        temperature = 0.3,
+        top_p= 0.95,
+        device=device
+    )
+    local_llm = HuggingFacePipeline(pipeline=pipe)
+    return local_llm
+
+
+def qa_llm():
+    llm = llm_pipeline()
+    embeddings = SentenceTransformerEmbeddings(model_name="all-MiniLM-L6-v2")
+    db = Chroma(persist_directory="vector_db_main", embedding_function = embeddings)
+    retriever = db.as_retriever()
+    qa = RetrievalQA.from_chain_type(
+        llm = llm,
+        chain_type = "stuff",
+        retriever = retriever,
+        return_source_documents=True
+    )
+    return qa
+
+def process_answer(instruction):
+    instruction = instruction
+    qa = qa_llm()
+    generated_text = qa(instruction)
+    answer = generated_text['result']
+    return answer
+
+def main():
+    # data_ingestion2()
+    print(process_answer("What is the EconVisor's motivation is deeply rooted in?"))
+
+if __name__ == "__main__":
+    main()
+
